Submit CustomerOrder partially works
We need to remove the constraint that only allows unique menu item IDs. If a unique menu item is selected, HTML page displays successfully. However, it displays the Customer object and MenuItem object instead of properties of the Customer/MenuItem objects. (We tried properties originally, like currentorder.customer.firstName, but this threw an error.)